backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
python-multipart==0.0.9
python-dotenv==1.0.0
aiohttp==3.9.5
jinja2==3.1.4
sqlalchemy==2.0.23
alembic==1.12.1
requests==2.31.0
pypdf==5.1.0
python-docx==1.1.0